refactor email service with transport strategies and templates
