+++ backend/src/auth/auth.module.ts
@@ -9,7 +9,7 @@ import { JwtStrategy } from './jwt.strategy';
 import { RefreshToken } from './refresh-token.entity';
 import { VerificationToken } from './verification-token.entity';
 import { User } from '../users/user.entity';
-import { EmailService } from '../common/email.service';
+import { EmailService } from '../common/email';
 import { Company } from '../companies/entities/company.entity';
 import { CompanyUser } from '../companies/entities/company-user.entity';
