+++ backend/src/common/email/index.ts
@@ -1,0 +1,3 @@
+export * from './email.service';
+export * from './transports';
+export * from './templates';
