+++ backend/src/common/email/templates/index.ts
@@ -1,0 +1,4 @@
+export * from './password-reset.mail';
+export * from './verification.mail';
+export * from './invitation.mail';
+export * from './added-to-company.mail';
