+++ backend/src/common/email/templates/password-reset.mail.ts
@@ -1,0 +1,10 @@
+import { SendMailOptions } from 'nodemailer';
+
+export function passwordResetMail(to: string, token: string): SendMailOptions {
+  return {
+    to,
+    subject: 'Password Reset Request',
+    text: `Your password reset token is: ${token}`,
+    html: `<p>Your password reset token is: <strong>${token}</strong></p>`,
+  };
+}
