+++ backend/src/common/email/transports/index.ts
@@ -1,0 +1,3 @@
+export * from './transport.interface';
+export * from './ethereal.transport';
+export * from './smtp.transport';
