+++ backend/src/companies/__tests__/invitations.accept-existing-user.spec.ts
@@ -6,7 +6,8 @@ import { Invitation, InvitationRole } from '../entities/invitation.entity';
 import { CompanyUser, CompanyUserRole } from '../entities/company-user.entity';
 import { Company } from '../entities/company.entity';
 import { User } from '../../users/user.entity';
-import { EmailService } from '../../common/email.service';
+import { EmailService } from '../../common/email';
+import { SendMailOptions } from 'nodemailer';
 
 describe('InvitationsService acceptExistingUser', () => {
   let service: InvitationsService;
@@ -16,12 +17,7 @@ describe('InvitationsService acceptExistingUser', () => {
   >;
   let usersRepo: jest.Mocked<Pick<Repository<User>, 'findOne' | 'save'>>;
   let companiesRepo: jest.Mocked<Pick<Repository<Company>, 'findOne'>>;
-  let emailService: {
-    sendAddedToCompanyEmail: jest.Mock<
-      void,
-      [string, string, InvitationRole]
-    >;
-  };
+  let emailService: { send: jest.Mock<void, [SendMailOptions]> };
 
   beforeEach(() => {
     invitationsRepo = {
@@ -40,13 +36,8 @@ describe('InvitationsService acceptExistingUser', () => {
       findOne: jest.fn(async () => Object.assign(new Company(), { id: 7, name: 'Co' })),
     } as unknown as jest.Mocked<Pick<Repository<Company>, 'findOne'>>;
     emailService = {
-      sendAddedToCompanyEmail: jest.fn(),
-    } as {
-      sendAddedToCompanyEmail: jest.Mock<
-        void,
-        [string, string, InvitationRole]
-      >;
-    };
+      send: jest.fn<void, [SendMailOptions]>(),
+    } as { send: jest.Mock<void, [SendMailOptions]> };
     service = new InvitationsService(
       invitationsRepo as unknown as Repository<Invitation>,
       companyUsersRepo as unknown as Repository<CompanyUser>,
@@ -85,11 +76,11 @@ describe('InvitationsService acceptExistingUser', () => {
       invitedBy: 1,
     });
     expect(invitation.acceptedAt).toBeInstanceOf(Date);
-    expect(emailService.sendAddedToCompanyEmail).toHaveBeenCalledWith(
-      'existing@user.com',
-      'Co',
-      InvitationRole.ADMIN,
-    );
+    const [[options]] = emailService.send.mock.calls;
+    expect(options.to).toBe('existing@user.com');
+    expect(options.subject).toBe('You were added to a company');
+    expect((options.html as string)).toContain('Co');
+    expect((options.html as string)).toContain('Admin');
   });
 
   it('rejects when email mismatch', async () => {
