+++ backend/src/companies/invitations.service.ts
@@ -21,7 +21,8 @@ import { User, UserRole } from '../users/user.entity';
 import { CreateInvitationDto } from './dto/create-invitation.dto';
 import { AcceptInvitationDto } from './dto/accept-invitation.dto';
 import * as crypto from 'crypto';
-import { EmailService } from '../common/email.service';
+import { EmailService } from '../common/email';
+import { invitationMail, addedToCompanyMail } from '../common/email/templates';
 import { validatePasswordStrength } from '../auth/password.util';
 
 @Injectable()
@@ -117,12 +118,14 @@ export class InvitationsService {
     const company = await this.companiesRepository.findOne({
       where: { id: companyId },
     });
-    await this.emailService.sendCompanyInvitationEmail(
-      dto.email,
-      rawToken,
-      company?.name ?? 'Your company',
-      dto.role,
-      invitation.expiresAt,
+    await this.emailService.send(
+      invitationMail(
+        dto.email,
+        rawToken,
+        company?.name ?? 'Your company',
+        dto.role,
+        invitation.expiresAt,
+      ),
     );
 
     return saved;
@@ -170,12 +173,14 @@ export class InvitationsService {
     invitation.tokenHash = tokenHash;
     invitation.expiresAt = new Date(Date.now() + 7 * 24 * 60 * 60 * 1000);
     await this.invitationsRepository.save(invitation);
-    await this.emailService.sendCompanyInvitationEmail(
-      invitation.email,
-      rawToken,
-      invitation.company?.name ?? 'Your company',
-      invitation.role,
-      invitation.expiresAt,
+    await this.emailService.send(
+      invitationMail(
+        invitation.email,
+        rawToken,
+        invitation.company?.name ?? 'Your company',
+        invitation.role,
+        invitation.expiresAt,
+      ),
     );
     return invitation;
   }
@@ -279,10 +284,12 @@ export class InvitationsService {
     const company = await this.companiesRepository.findOne({
       where: { id: invitation.companyId },
     });
-    await this.emailService.sendAddedToCompanyEmail(
-      user.email,
-      company?.name ?? 'Your company',
-      invitation.role,
+    await this.emailService.send(
+      addedToCompanyMail(
+        user.email,
+        company?.name ?? 'Your company',
+        invitation.role,
+      ),
     );
 
     this.acceptAttempts.delete(tokenHash);
@@ -353,10 +360,12 @@ export class InvitationsService {
     const company = await this.companiesRepository.findOne({
       where: { id: invitation.companyId },
     });
-    await this.emailService.sendAddedToCompanyEmail(
-      savedUser.email,
-      company?.name ?? 'Your company',
-      invitation.role,
+    await this.emailService.send(
+      addedToCompanyMail(
+        savedUser.email,
+        company?.name ?? 'Your company',
+        invitation.role,
+      ),
     );
 
     this.acceptAttempts.delete(tokenHash);
