+++ backend/src/users/users.service.ts
@@ -7,7 +7,8 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { FindOptionsWhere, MoreThan, Repository } from 'typeorm';
 import * as crypto from 'crypto';
 
-import { EmailService } from '../common/email.service';
+import { EmailService } from '../common/email';
+import { passwordResetMail } from '../common/email/templates';
 
 import { User, UserRole } from './user.entity';
 import { CreateUserDto } from './dto/create-user.dto';
@@ -63,7 +64,7 @@ export class UsersService {
     user.passwordResetToken = hashedToken;
     user.passwordResetExpires = new Date(Date.now() + 60 * 60 * 1000);
     await this.usersRepository.save(user);
-    await this.emailService.sendPasswordResetEmail(user.email, token);
+    await this.emailService.send(passwordResetMail(user.email, token));
   }
 
   async resetPassword(token: string, password: string): Promise<User> {
